+++ docs/changelog.rst
@@ -20,6 +20,11 @@ New library methods:
 
 -  ``package_records``
 
+Changed
+~~~~~~~
+
+- mapping-sheet: Improved documentation of ``--extension`` and ``--extension-field``.
+
 0.1.3 (2019-09-26)
 ------------------